Rename `instance` to `tmpl`
This is something that has bugged me for a while in various Meteor tutorials available online. Now that we're working in an open forum of best practices I'd like to put forward the case that, in short, naming a variable `instance` is just as non-expressive as declaring one called "variable" (`var variable = xyz`).

The fact that it comes from `Template.instance()` is no reason to call it `instance` – an equivalent (and equivalently) poor case would be `var instance = new TodoItem()`. So, in the same way that we declare `var todoItem = new TodoItem()`, which is considered good practice in both the JavaScript world and beyond, the instance of a Template should be called `template`, or, in short-hand `tmpl`. By lucky coincidence, this abbreviation is as concise (with the same number of letters) as `this`, which it (in many cases) represents.

This change encourages good naming practices for newcomers to Javascript, and improves readability and clarity for more advanced developers who are new to Meteor.

-----

I also fixed a double "this.this" which is invalid Javascript.