Some additions to the Meteor/React guide
Open to discussion of course! People might want to learn more
about Redux, what it is, if they need it, what advantages it
brings to the table. Performance testing is a biggie, because
components can often go on a re-render binge and cause problems.
These tools can help measure and pinpoint trouble spots.