Testing: Replace `factory` to `dburles:factory`
In the testing section of the guide (amongst others), the guide references the Meteor "Todos" app: https://github.com/meteor/todos

Previously, `factory` was a locally forked Meteor package within the Todos repo which was a fork of `dburles:factory` with (minimal) tweaks.  See discussion in:

* meteor/todos#193
* meteor/todos#96

With the landing of meteor/todos#202, the Todos application now utilizes `dburles:factory`.  This commit gives the same treatment as that PR provided to `meteor/todos` (https://github.com/meteor/todos/pull/202/files).

Closes meteor/guide#576