SSL: Clarify the situations where `force-ssl` should be used.
Previously, the `force-ssl` package was a recommendation for providing the redirect from HTTP to HTTPS, but this is no longer the recommended approach.  Now, the recommendation is to provide this redirection on the proxy server (or provider) which terminates the SSL.  The `force-ssl` package can still be used in cases where this is not possible, but should be used as a last resort.

Relates to meteor/meteor#8268