Updated loading data with Meteor Methods
I wanted to add a blurb about loading data with Meteor Methods. It should be noted that loading the _same_ data with Meteor Method also does not share publications, observers, and data between clients how pub/sub can and that it's super inefficient on the database.  I just learned this the hard way.